% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTables.R
\name{createTables}
\alias{createTables}
\title{Creates metadata tables}
\usage{
createTables(connectionDetails, resultsDatabaseSchema,
  outputFolder = "output")
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} 
in the \code{DatabaseConnector} package.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. 
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{outputFolder}{A folder to hold the logs.}
}
\value{
Nothing
}
\description{
\code{CdmMetadata} creates metadata tables within a results database schema.
}
\details{
\code{CdmMetadata} creates metadata tables within a results database schema.
}
