% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KronosInsertMetadata.R
\name{insertKronosMetadata}
\alias{insertKronosMetadata}
\title{Insert Kronos Results into Metadata Tables.}
\usage{
insertMetadata(achillesResultSet)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{vocabDatabaseSchema}{The schema that contains the OMOP vocab tables}

\item{resultsDatabaseSchema}{The schema that contains the metadata table}

\item{rdsDir}{(Optional) a directory to store the results as RDS files}
}
\description{
Build monthly time series objects using \cr\code{buildAchillesTs} on data generated by \cr\code{getAchillesData}
            and insert relevant time series analyses into metadata tables.
}
\details{
This function will parse a data frame into time series objects, run analyses, then insert into metadata tables.
        The dataframe is assumed to come from \cr\code{getAchillesData} and have the correct structure.
}
