% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaunchShinyApps.R
\name{launchMaintenanceApp}
\alias{launchMaintenanceApp}
\title{Launches the Maintenance Shiny App}
\usage{
launchMaintenanceApp(connectionDetails, cdmDatabaseSchema,
  resultsDatabaseSchema, vocabDatabaseSchema = cdmDatabaseSchema)
}
\arguments{
\item{vocabDatabaseSchema}{}
}
\description{
Launches the Maintenance Shiny App
}
