% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaunchShinyApps.R
\name{launchMetadataApp}
\alias{launchMetadataApp}
\title{Launches the Shiny Metadata App}
\usage{
launchMetadataApp(jsonPath)
}
\arguments{
\item{jsonPath}{The path to the Sources JSON file}
}
\description{
Launches the Shiny Metadata App
}
